join support in verify()
join directive loads corresponding records from another table, for example, for a rule aboud USBD, you can join the corresponding USOMVS record, and test if an ID has UID(0)

{'specials must not have root':
                        (['USBD'],
                         {'id': '101',
                          'join': 'USOMVS',
                          'find': {'field':'SPECIAL','value':'YES'},
                          'test': [{'field':'UID', 'value':'0000000000','action':'FAIL'}],
                         }
                        )
                      }

default join is by index field(s), but this can be overridden using

'join': {'table':'GPOMVS', 'on':'USBD_DEFGRP_ID'}

join can also refer to connectData, in which case the index of the destination (connects) is determined from the type of the join source: you get to check "your" connects.

{'specials must be connected to SYS1, not as group special and not revoked':
                        (['USBD'],
                         {'id': '103',
                          'join': {'table':'USCON','how':'left'},
                          'find': [{'field':'SPECIAL','value':'YES'},
                                   {'field':'GRP_ID','value':'SYS1'}],
                          'test': [{'field':'GRP_SPECIAL', 'value':'NO'},
                                   {'field':'USCON_REVOKE', 'value':'NO'}]
                         }
                        )
                      }

default join method is 'inner', so records will only be verified if matches exists in both tables.  This can be modified with 'on':'left'

field names from the joined table can be used in 'find' 'skip' and 'test' field rereferences, with or without field prefix.

'find' and 'skip' are performed twice: once for the fields in the base (subjectDF) table before join, so as to reduce the number of records.
again after join and match  action (if any and if needed!) to take into account fields from those sources.

test results can be reversed.  by default, you specify how you expect a field value to be, if the expected value is found, the entry is compliant and will not be shown.
add  'action':'failure' in you test directive to reverse the direction (see 1st example)

default domain name SPECIAL (referring to *, &RACUID and RACGID) renamed to SPECIALID, aligning with ACLID

nameInColumns: utility to check if a column name without a prefix matches any column name